Implement new track and cell. Tweak ITC_DEFER macro. Implement Plug::DisconnectAll() which I apparently forgot about earlier. Get everything compiling again. A few style tweaks might have slipped in too; not sure.

+++ trunk/src/track.h
@@ -24,7 +24,7 @@ struct Track
 	void AddCell();
 
 	// Remove the cell at this index from this track.
-	void RemoveCell(unsigned int idx);
+	void RemoveCell(unsigned short int idx);
 
 	// Render for this much time since the last pass.
 	void R_Render(float time);
